geiger/requirements.txt
numpy
pyzmq
matplotlib